refactor: update generic dialog to have title and maybe description
